adding crud

+++ src/main/java/tn/esprit/immobilier/repositories/IAttachmentRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.immobilier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.immobilier.entities.Attachement;
 
-public interface IAttachmentRepository extends JpaRepository<Attachement, Integer> {
+public interface IAttachmentRepository extends JpaRepository<Attachement, Long> {
 }
