+++ src/main/java/tn/esprit/immobilier/repositories/IImmobilierRepository.java
@@ -2,6 +2,7 @@ package tn.esprit.immobilier.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.immobilier.entities.Immobilier;
+import tn.esprit.immobilier.entities.Jeton;
 
-public interface IImmobilierRepository extends JpaRepository<Immobilier, Integer> {
+public interface IImmobilierRepository  extends JpaRepository<Immobilier,Long> {
 }
