+++ src/main/java/tn/esprit/immobilier/repositories/IPositionRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.immobilier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.immobilier.entities.Position;
 
-public interface IPositionRepository extends JpaRepository<Position, Integer> {
+public interface IPositionRepository extends JpaRepository<Position, Long> {
 }
