+++ src/main/java/tn/esprit/immobilier/repositories/IReservationRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.immobilier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.immobilier.entities.Reservation;
 
-public interface IReservationRepository extends JpaRepository<Reservation, Integer> {
+public interface IReservationRepository extends JpaRepository<Reservation, Long> {
 }
